Adding updated networkx graph for time-series
