+++ pywhy_graphs/classes/__init__.py
@@ -1,10 +1,8 @@
 from .admg import ADMG
 from .cpdag import CPDAG
 from .pag import PAG
-from .timeseries import (
+from .timeseries.timeseries import (
     StationaryTimeSeriesDiGraph,
     StationaryTimeSeriesGraph,
     StationaryTimeSeriesMixedEdgeGraph,
-    TimeSeriesDiGraph,
-    TimeSeriesGraph,
 )
